requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
cloudinary==1.33.0
dj3-cloudinary-storage==0.0.6
Django==4.2.3
django-ckeditor==6.6.1
django-cors-headers==4.2.0
django-js-asset==2.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.0.1
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
gunicorn==20.1.0
idna==3.4
Pillow==10.0.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
PyJWT==2.6.0